working on the server story before release
